Включение динамической компиляции и добавление нового контроллера
