+++ WebShop/Startup.cs
@@ -17,7 +17,7 @@ namespace WebShop
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllersWithViews();
+            services.AddControllersWithViews().AddRazorRuntimeCompilation();
         }
