+++ WebShop/Views/Employees/Details.cshtml
@@ -1,12 +1,5 @@
 ﻿@model WebShop.Models.Employee
 
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>WebStore - Сотрудник @Model.Surname @Model.Name @Model.Patronymic</title>
-</head>
 <body>
 
 <h3>Сотрудник</h3>
@@ -27,6 +20,5 @@
     <dd>@Model.Age</dd>
 </dl>
 <hr/>    
-    <a href="~/Home/Employees">К списку</a>
+<a href="~/Home/Employees">К списку</a>
 </body>
-</html>
