+++ WebShop/Views/Employees/Index.cshtml
@@ -1,0 +1,35 @@
+﻿@model IEnumerable<WebShop.Models.Employee>
+
+<body>
+
+<h3>Сотрудники</h3>
+<hr/>
+
+<table class="table">
+    <thead>
+    <tr>
+        <th>Id</th>
+        <th>Фамилия</th>
+        <th>Имя</th>
+        <th>Отчество</th>
+        <th>Возраст</th>
+        <th></th>
+    </tr>
+    </thead>
+    <tbody>
+    @foreach (var employee in Model)
+    {
+        <tr>
+            <td>@employee.Id</td>
+            <td>@employee.Surname</td>
+            <td>@employee.Name</td>
+            <td>@employee.Patronymic</td>
+            <td>@employee.Age</td>
+            <td><a href="~/Home/EmployeeInfo/@employee.Id">Информация</a></td>
+        </tr>
+    }
+    </tbody>
+</table>
+<hr/>
+<a href="~/">На главную</a>
+</body>
