--- WebShop/Views/Home/Employees.cshtml
@@ -1,43 +1,0 @@
-﻿@model IEnumerable<WebShop.Models.Employee>
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>WebStore - @ViewBag.Title</title>
-</head>
-<body>
-
-<h3>Сотрудники</h3>
-<hr/>
-
-<table>
-    <thead>
-    <tr>
-        <th>Id</th>
-        <th>Фамилия</th>
-        <th>Имя</th>
-        <th>Отчество</th>
-        <th>Возраст</th>
-        <th></th>
-    </tr>
-    </thead>
-    <tbody>
-        @foreach (var employee in Model)
-        {
-            <tr>
-                <td>@employee.Id</td>
-                <td>@employee.Surname</td>
-                <td>@employee.Name</td>
-                <td>@employee.Patronymic</td>
-                <td>@employee.Age</td>
-                <td><a href="~/Home/EmployeeInfo/@employee.Id">Информация</a></td>
-            </tr>
-        }
-    </tbody>
-</table>
-<hr/>
-<a href="~/">На главную</a>
-</body>
-</html>
